requirements.txt
# Python dependencies for Robot Card Generator
# Install with: pip install -r requirements.txt

jinja2>=3.1.0
pyyaml>=6.0
